app/requirements/requirements.txt
caribou
Flask
holidays
python-dateutil
requests
six
tendo
urllib3
paho-mqtt
uwsgi